Unit tests for urb module and working mock solution, fixes #7
