+++ modules/resolver.go
@@ -1,5 +1,7 @@
 package modules
 
+import "self-stabilizing-uniform-reliable-broadcast/constants"
+
 // ModuleType indicates the type of module
 type ModuleType int
 
@@ -12,6 +14,13 @@ const (
 	THETAFD ModuleType = 2
 )
 
+// IResolver defines what interface functions are available for inter-module communication
+// Also makes it testable..
+type IResolver interface {
+	hb() []int
+	trusted() []int
+}
+
 // Resolver facilitates inter-module communication
 type Resolver struct {
 	Modules map[ModuleType]interface{}
@@ -19,10 +28,17 @@ type Resolver struct {
 
 func (r *Resolver) hb() []int {
 	m := r.Modules[HBFD].(HbfdModule)
-	return m.hb()
+	return m.Hb
 }
 
 func (r *Resolver) trusted() []int {
 	m := r.Modules[THETAFD].(ThetafdModule)
-	return m.trusted()
+	trusted := []int{}
+	for idx, x := range m.Vector {
+		if x < constants.THETAFD_W {
+			trusted = append(trusted, idx)
+		}
+	}
+
+	return trusted
 }
